requirements.txt
streamlit 
groq
pillow
requests
python-dotenv
pdf2image